added unique tag tracking
loads the unique tags within that exists in the flashcards. Plan to use the list to make a Select/De-Select UI element rather than a input box that could lead to multiple versions of the same tag. Also to eventually add filtering cards by tags easily on the practice screen.